pyproject.toml
[tool.poetry]
name = "gymfreak-diary"
version = "0.1.0"
description = "Simple application to note down your workouts details"
authors = ["jp <plutakuba@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
pydantic-settings = "^2.0.3"
pydantic = {extras = ["email"], version = "^2.4.2"}
psycopg2 = "^2.9.9"
pathlib = "^1.0.1"
pandas = "^2.1.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
httpx = "^0.25.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
pre-commit = "^3.5.0"
isort = "^5.12.0"
pylint = "^3.0.2"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff]
extend-exclude = ["tests", "migrations/*", "__init__.py"]

[tool.codespell]
skip = "migrations/**/*,migrations,migrations/*.py"




